requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
boto3==1.35.76
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
motor==3.6.0
python-socketio==5.11.0
ollama==0.4.4
requests==2.32.3
python-jose[cryptography]==3.3.0

# RAG Pipeline
chromadb==0.4.22
# pypdf2==3.0.1
pypdf==5.1.0
python-multipart==0.0.6